[1.0.0] Only pass clean user data instead of raw user data to views. Additional: Fixed scroll chat.
