+++ src/config/passport/strategy/index.js
@@ -1,0 +1,3 @@
+export { default as facebookStrategy } from "./facebookStrategy";
+export { default as googleStrategy } from "./googleStrategy";
+export { default as localStrategy } from "./localStrategy";
