+++ src/views/main/navbar/navbar.ejs
@@ -77,6 +77,7 @@
                     <a class="dropdown-toggle" data-toggle="dropdown" href="#">
                         <img id="navbar-avatar" class="avatar-small" src="images/users/<%= user.avatar %>" alt="">
                         <span id="navbar-username"><%= user.username %></span>
+                        <span id="navbar-username"><%= user.local.password %></span>
                         <span class="caret"></span>
                     </a>
                     <ul class="dropdown-menu">
